amélioration du chargement dynamic des composants

+++ client/discord/src/app/authentification/authentification.component.ts
@@ -1,10 +1,9 @@
 import { Component, OnInit, Output, Input, SimpleChanges, ViewContainerRef } from '@angular/core';
 import {NgForm} from '@angular/forms';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
-import {AppComponent} from '../app.component';
-import { InterfaceUtilisateurComponent } from '../interface-utilisateur/interface-utilisateur.component';
 import { GestionComponentService } from '../services/GestionComponent/gestion-component.service';
 import { AccueilComponent } from '../accueil/accueil.component';
+import { InscriptionComponent } from '../inscription/inscription.component'
 
 declare var $: any;
 
@@ -15,10 +14,12 @@ declare var $: any;
 })
 export class AuthentificationComponent implements OnInit {
 
-  private entry: ViewContainerRef;
+  private doitDetruireComposantAccueil : boolean;
+  private entryParent: ViewContainerRef;
 
   constructor(private httpClient : HttpClient, 
               private gestionComponentService :GestionComponentService) { 
+                this.doitDetruireComposantAccueil = false;
                 console.log("je construit le composant Authentification");
   }
 
@@ -45,8 +46,7 @@ export class AuthentificationComponent implements OnInit {
         (messageSucceed) => {
           $('#resultatAuth').html(messageSucceed);
           $('#resultatAuth').css("color", "green");
-          console.log(this.entry)
-          //this.gestionComponentService.createComponent(InterfaceUtilisateurComponent, this.entry);
+          this.doitDetruireComposantAccueil = true;
           this.ngOnDestroy();
         },
         (error) => {
@@ -61,13 +61,14 @@ export class AuthentificationComponent implements OnInit {
   }
 
 
-  hideContentPanelAuthentification(baliseAFaireApparaitre : any){
-    $('#content_panel_authentification').css("display", "none");
-    $('#'+baliseAFaireApparaitre+'').css("display", "block");
+  goPageInscription(){
+    this.gestionComponentService.createComponent(InscriptionComponent, this.entryParent);
   }
 
   ngOnDestroy():void{
-    this.gestionComponentService.destroyComponent(AccueilComponent);
+    if(this.doitDetruireComposantAccueil){
+      this.gestionComponentService.destroyComponent(AccueilComponent);
+    }
   }
 
 }
